[Common-TextInput] Moved ElementWrapper to elements folder and other formatting changes.
