+++ source/reactnative/src/components/containers/column.js
@@ -10,13 +10,14 @@ import {
 	StyleSheet, 
 	Dimensions 
 } from 'react-native';
+
 import { Registry } from '../registration/registry'
 import * as Utils from '../../utils/util';
 import * as Constants from '../../utils/constants';
 import * as Enums from '../../utils/enums';
 import { SelectAction } from '../actions';
-import { HostConfigManager } from '../../utils/host-config'
-import { StyleManager } from '../../styles/style-config'
+import { HostConfigManager } from '../../utils/host-config';
+import { StyleManager } from '../../styles/style-config';
 
 const deviceWidth = Dimensions.get('window').width;
