+++ source/reactnative/src/components/elements/text-block.js
@@ -10,14 +10,14 @@ import {
     Text
 } from 'react-native';
 
-import { HostConfigManager } from "../../utils/host-config";
+import { HostConfigManager } from '../../utils/host-config';
 import { TextFormatter } from '../../utils/text-formatters';
-import ElementWrapper from '../inputs/element-wrapper';
+import ElementWrapper from '../elements/element-wrapper';
 import * as Utils from '../../utils/util';
 import * as Enums from '../../utils/enums';
 import * as Constants from '../../utils/constants';
 import processMDText from '../../utils/formatText';
-import { StyleManager } from '../../styles/style-config'
+import { StyleManager } from '../../styles/style-config';
 
 
 export class TextBlock extends React.Component {
