+++ source/reactnative/src/components/inputs/index.js
@@ -3,5 +3,5 @@ export * from './number-input';
 export * from './toggle-input';
 export * from './date-input';
 export * from './time-input';
-export * from './element-wrapper';
+export * from './picker-input';
 export * from './choice-set/choice-set-input';
